snapmbed/requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
piexif>=1.1.3
mutagen>=1.47.0
pillow-heif>=0.13.0
pyinstaller>=6.0.0
